Renamed the path of accessList and accessTuple
